refactor: extract JaCoCo parser to shared module + Java coverage (10k.1.2)
CLAUDE.md rule #2 ("Each tool has ONE gather function") — JaCoCo
support was inline in src/languages/kotlin.ts, but Java needs the
SAME parser. Forking would have been the architectural violation.
Best-case SSOT validation:

- The parser was ALREADY generic. kotlin's existing tests cover it
  against `jacoco-java-source.xml` (line 149 of languages-kotlin.test.ts
  pre-refactor) — kotlin's tests literally validate Java-source
  JaCoCo XML. The parser had no kt-specific strings.
- The gather glue was identical (find report path, read file, call
  parser, wrap in CoverageResult envelope). Both packs need exactly
  this; no language-specific dispatch.

Shared module: `src/analyzers/tools/jacoco.ts` — exports:
- `parseJaCoCoXml(raw, sourceFile, cwd)` — language-agnostic XML parser
- `findJaCoCoReport(cwd)` — checks Gradle (Android, plain JVM) +
  MAVEN paths (target/site/jacoco/jacoco.xml, target/site/
  jacoco-aggregate/jacoco.xml — both new for Java) + fallbacks
- `gatherJaCoCoCoverageResult(cwd)` — the identical glue, now called
  by both packs' provider.gather

Kotlin pack: dropped 130+ lines of inline JaCoCo logic.
`kotlinCoverageProvider.gather` is a 1-line delegate now. Public
`parseJaCoCoXml` re-export removed; test/languages-kotlin.test.ts
imports from the shared module instead.

Java pack: `javaCoverageProvider` registered. 2 new tests exercise
the gather glue against the existing Java-source JaCoCo fixture
laid out under `target/site/jacoco/jacoco.xml` (the Maven default
path that's new for Java).

Recipe v3 reflection: BEST CASE for SSOT. Architectural debt was
already paid (parser was generic from day 1); the refactor was just
relocating to the right home. No new Recipe v3 gaps surfaced. The
"reuse parser unchanged" test from the 10k roadmap held — kotlin's
parser works for Java without modification.

Tests: 941 passed | 2 skipped (52/52 files), +2 from 10k.1.1.
Capabilities pipeline: 3 of 5 done (testFramework + imports +
coverage). Next: 10k.1.3 PMD lint.
