+++ src/languages/kotlin.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 
-import { type Coverage, type FileCoverage, round1 } from '../analyzers/tools/coverage';
+import { gatherJaCoCoCoverageResult } from '../analyzers/tools/jacoco';
 import { getFindExcludeFlags } from '../analyzers/tools/exclusions';
 import {
   classifyOsvSeverity,
@@ -203,154 +203,15 @@ const kotlinLintProvider: CapabilityProvider<LintResult> = {
 };
 
 // ─── Coverage (JaCoCo XML) ──────────────────────────────────────────────────
-
-/**
- * Pure parser for JaCoCo's XML report (DTD: `report.dtd`). The structure:
- *
- *   <report name="...">
- *     <package name="com/example">
- *       <class name="com/example/Foo" sourcefilename="Foo.kt">...</class>
- *       <sourcefile name="Foo.kt">
- *         <line nr="N" mi="..." ci="..." mb="..." cb="..."/>
- *         <counter type="LINE" missed="X" covered="Y"/>
- *       </sourcefile>
- *       <counter type="LINE" missed="X" covered="Y"/>
- *     </package>
- *     <counter type="LINE" missed="X" covered="Y"/>
- *   </report>
- *
- * Per-file coverage comes from `<sourcefile>` blocks: their LINE counter
- * holds the file's missed/covered totals. Project-level total comes from
- * the top-level `<counter type="LINE">` (last in the document).
- *
- * Path attribution joins `<package name>` (forward-slashed, JVM-style)
- * with `<sourcefile name>` to produce the canonical relative path the
- * downstream consumers expect (`com/example/Foo.kt`). JVM bytecode
- * namespacing isn't 1:1 with on-disk source paths in multi-module
- * projects — accepted limitation (matches C#'s cobertura attribution).
- *
- * Returns null when no `<counter type="LINE">` exists at the top level
- * — that's JaCoCo's "no coverage data" signal, distinct from "0%
- * coverage" (where the counter exists with covered=0).
- *
- * Exported for unit tests; consumed by `gatherKotlinCoverageResult`.
- */
-export function parseJaCoCoXml(raw: string, sourceFile: string, _cwd: string): Coverage | null {
-  const files = new Map<string, FileCoverage>();
-  // Iterate <package> blocks. Each block contains <sourcefile> children
-  // we tally and aggregate counters we can ignore (they're sums of the
-  // children, redundant given we sum the children ourselves).
-  const packageRe = /<package\s+name="([^"]+)">([\s\S]*?)<\/package>/g;
-  let pm: RegExpExecArray | null;
-  while ((pm = packageRe.exec(raw)) !== null) {
-    const pkgPath = pm[1].replace(/\\/g, '/'); // JVM uses forward-slashes already; defensive
-    const pkgInner = pm[2];
-    // Within a <package>, <sourcefile> blocks own per-file counters.
-    const sourceFileRe = /<sourcefile\s+name="([^"]+)">([\s\S]*?)<\/sourcefile>/g;
-    let sm: RegExpExecArray | null;
-    while ((sm = sourceFileRe.exec(pkgInner)) !== null) {
-      const fileName = sm[1];
-      const sourceInner = sm[2];
-      // The LAST <counter type="LINE"> in <sourcefile> is the
-      // file-level aggregate; earlier ones are per-method. We pick the
-      // last to get the full-file roll-up.
-      const counterRe = /<counter\s+type="LINE"\s+missed="(\d+)"\s+covered="(\d+)"\s*\/>/g;
-      let lastMissed = 0;
-      let lastCovered = 0;
-      let cm: RegExpExecArray | null;
-      while ((cm = counterRe.exec(sourceInner)) !== null) {
-        lastMissed = parseInt(cm[1], 10);
-        lastCovered = parseInt(cm[2], 10);
-      }
-      const total = lastMissed + lastCovered;
-      const rel = pkgPath ? `${pkgPath}/${fileName}` : fileName;
-      files.set(rel, {
-        path: rel,
-        covered: lastCovered,
-        total,
-        pct: round1(total > 0 ? (lastCovered / total) * 100 : 0),
-      });
-    }
-  }
-
-  // Top-level project-wide LINE counter — JaCoCo emits it after the
-  // last </package>. Use a non-greedy match against the document tail
-  // to avoid grabbing per-package counters as project-level.
-  const tailMatch = raw.match(
-    /<\/package>\s*<counter\s+type="LINE"\s+missed="(\d+)"\s+covered="(\d+)"\s*\/>\s*<\/report>/,
-  );
-  let totalMissed = 0;
-  let totalCovered = 0;
-  if (tailMatch) {
-    totalMissed = parseInt(tailMatch[1], 10);
-    totalCovered = parseInt(tailMatch[2], 10);
-  } else {
-    // No project-level counter (degenerate report — single package, no
-    // explicit roll-up). Sum the per-file totals as the linePercent
-    // basis.
-    for (const f of files.values()) {
-      totalCovered += f.covered;
-      totalMissed += f.total - f.covered;
-    }
-  }
-  const grandTotal = totalCovered + totalMissed;
-  if (grandTotal === 0 && files.size === 0) return null;
-
-  return {
-    source: 'jacoco',
-    sourceFile,
-    linePercent: round1(grandTotal > 0 ? (totalCovered / grandTotal) * 100 : 0),
-    files,
-  };
-}
-
-/**
- * Standard JaCoCo report locations across project layouts:
- *   - app/build/reports/jacoco/jacocoTestReport/jacocoTestReport.xml
- *     (Android default — `app` module, `jacocoTestReport` task)
- *   - build/reports/jacoco/test/jacocoTestReport.xml
- *     (plain JVM Kotlin via the `jacoco` plugin's default `test` task)
- *   - build/reports/jacoco/jacocoTestReport/jacocoTestReport.xml
- *     (multi-module aggregate; many builds rename the task)
- *   - jacocoTestReport.xml (top-level — fixture / direct path)
- *
- * Multi-module projects emit per-module reports under `<module>/build/...`;
- * the cross-ecosystem fixture in step 5 will calibrate which paths show
- * up reliably. Until then, the list above is conservative-priority: most
- * specific (app-module) first, root fallback last.
- */
-function findJaCoCoReport(cwd: string): string | null {
-  const candidates = [
-    'app/build/reports/jacoco/jacocoTestReport/jacocoTestReport.xml',
-    'build/reports/jacoco/test/jacocoTestReport.xml',
-    'build/reports/jacoco/jacocoTestReport/jacocoTestReport.xml',
-    'jacocoTestReport.xml',
-  ];
-  for (const rel of candidates) {
-    const abs = path.join(cwd, rel);
-    if (fs.existsSync(abs)) return rel;
-  }
-  return null;
-}
-
-function gatherKotlinCoverageResult(cwd: string): CoverageResult | null {
-  const reportRel = findJaCoCoReport(cwd);
-  if (!reportRel) return null;
-  let raw: string;
-  try {
-    raw = fs.readFileSync(path.join(cwd, reportRel), 'utf-8');
-  } catch {
-    return null;
-  }
-  const coverage = parseJaCoCoXml(raw, reportRel, cwd);
-  if (!coverage) return null;
-  return { schemaVersion: 1, tool: `coverage:${coverage.source}`, coverage };
-}
+//
+// Parser, finder, and gather glue all live in `src/analyzers/tools/jacoco.ts`
+// — language-agnostic SSOT (CLAUDE.md rule #2). The kotlin pack just
+// declares its provider and delegates. Java pack does the same.
 
 const kotlinCoverageProvider: CapabilityProvider<CoverageResult> = {
   source: 'kotlin',
   async gather(cwd) {
-    return gatherKotlinCoverageResult(cwd);
+    return gatherJaCoCoCoverageResult(cwd);
   },
 };
